added confirmation dialog for deleting books

+++ src/app/pages/book/book.component.ts
@@ -2,6 +2,7 @@ import { HttpClient } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Book } from 'src/app/interfaces/Book';
+import { LanguagePipe } from 'src/app/pipes/language.pipe';
 import { AuthenticationService } from 'src/app/services/authentication.service';
 import { LanguageService } from 'src/app/services/language.service';
 
@@ -17,6 +18,7 @@ export class BookComponent implements OnInit {
 
   constructor(
     public languageService: LanguageService,
+    private languagePipe: LanguagePipe,
     public authenticationService: AuthenticationService,
     private route: ActivatedRoute,
     private http: HttpClient,
@@ -30,6 +32,9 @@ export class BookComponent implements OnInit {
   }
 
   handleDelete(): void {
+    if (!window.confirm(this.languagePipe.transform('deleteConfirmation'))) {
+      return;
+    }
     this.http.delete(`/books/${this.bookId}`)
       .subscribe(() => this.router.navigate(['/books']));
   }
